requirements.txt
streamlit
opencv-python
numpy>=1.26.0
Pillow
torch
torchvision
flask
flask-cors
python-dotenv
twilio
requests
